Release 2.0.1

+++ src/main/java/com/tibco/sonar/plugins/bw5/sensor/ProcessMetricSensor.java
@@ -55,7 +55,7 @@ public class ProcessMetricSensor implements Sensor {
 
     }
 
-    private void saveMeasure(Metric metric, double value, InputFile processFileResource) {
+    private void saveMeasure(Metric metric, int value, InputFile processFileResource) {
         sensorContext.<Integer>newMeasure()
                 .forMetric(metric)
                 .on(processFileResource)
